// src/app/modules/browse/models/user.ts
export interface User {
    "IdUser": number
    "FirstName": string
    "LastName": string
    "Login": string
    "Password": string
    "Salt": string
    "WebSite": string
    "Twitter": string
    "Instagram": string
    "Facebook": string
    "Email": string
    "Phone": string
    "Mobile": string
    "Country": string
    "City": number
    "Picture": string
}